requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.7
pymongo==4.5.0
sqlalchemy==1.4.49
numpy==1.24.3
python-dotenv==1.0.0
pyspark==3.4.1
dbt-core==1.6.2
dbt-postgres==1.6.2
